+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Yii Framework 2 elasticsearch extension Change Log
 - Enh #43: Elasticsearch log target (trntv, beowulfenator)
 - Bug: Added accessor method for the default elasticsearch primary key (kyle-mccarthy)
 - Enh #119: Added support for explanation on query (kyle-mccarthy)
+- Enh #150: Explicitily send `Content-Type` header in HTTP requests to elasticsearch (lubobill1990)
 - Enh: Deserialization of raw response when text/plain is supported (Tezd)
 - Enh: Added ability to work with aliases through Command class (Tezd)