+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Yii Framework 2 elasticsearch extension Change Log
 2.0.5 under development
 -----------------------
 
+- Bug #161: Changed yii\base\Object to yii\base\BaseObject (sashsvamir)
 - Bug #120: Fix debug panel markup to be compatible with Yii 2.0.10 (drdim)
 - Bug: Updated debug panel classes to be consistent with yii 2.0.7 (beowulfenator)
 - Enh #15: Special data provider `yii\elasticsearch\ActiveDataProvider` created (klimov-paul)